requirements.txt
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.4.4
defusedxml==0.7.0rc2
Django==2.2
django-cleanup==5.1.0
django-mathfilters==1.0.0
django-messages==0.6.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-paypal==1.0.0
django-request==1.5.6
djangorestframework==3.12.2
forex-python==1.5
idna==2.10
install==1.3.4
jsonfield==3.1.0
oauthlib==3.1.0
paypal-payouts-sdk==1.0.0
paypalhttp==1.0.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
swapper==1.1.2.post1
urllib3==1.26.3
